+++ packages/webpack-config/src/loaders/index.ts
@@ -9,4 +9,3 @@ export {
 export { default as createBabelLoader } from './createBabelLoader';
 export * from './createBabelLoader';
 export { default as createFontLoader } from './createFontLoader';
-export { default as createWorkerLoader } from './createWorkerLoader';